pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = "A discord bot to manage bits discord server"
authors = ["Sattwik <hellosattwik@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.14"


[tool.poetry.group.bot.dependencies]
discord-py = "^2.3.2"


[tool.poetry.group.api.dependencies]
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-dracula-theme = "^1.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
